♻️ [REFACTOR] 연습 모드 구조 변경
